+++ lib/celluloid/io.rb
@@ -28,6 +28,13 @@ module Celluloid
       actor && actor.mailbox.is_a?(Celluloid::IO::Mailbox)
     end
 
+    def self.copy_stream(src, dst, copy_length = nil, src_offset = nil)
+      src = ::IO.try_convert(src)
+      dst = ::IO.try_convert(dst)
+
+      Celluloid.defer { ::IO.copy_stream(src, dst, copy_length, src_offset) }
+    end
+
     def wait_readable(io)
       io = io.to_io
       if IO.evented?